Carousel and Timeline
Changed how the project carousel is sized, it now takes up half the size of the webpage horizontially, this makes it fit the 16:9 image better than an entire section by itself. Also removed the place where a description of the carousel was going to go because it feels pretty self-explanatory how to use it.

Added a timeline to show work history. It is dynamically sized to be the same height as the carousel section and the same width